==> scripts/netmums/1.0-Create_Data-Scrape.ipynb <==
# Scrape Forum Data
This notebook scrapes the Netmums and inserts the results into a SQLite database

## Data Sources
- Netmums forum: https://www.netmums.com/coffeehouse/

## Changes
- 2020-01-30: Created
- 2020-02-02: Added thread scraping
- 2020-02-05: Moved functions to netmums.py

## TODO
- 

## Imports

import sqlite3
from pathlib import Path
from netmums import *
from scraping import *
from datetime import datetime
import os
import csv
import re
from tqdm import tqdm

## File Locations

p = Path.cwd()
path_parent = p.parents[0]

path_db = str(path_parent / "database" / "netmums-merged.db")
path_groups = str(path_parent / "scripts" / "netmums-group_{}.py")

## Get Forum List

Create list of Netmum forum pages in the db. Skip if already created. 

conn = create_connection(path_db)
cur = conn.cursor()

cur.execute(''' SELECT count(name) FROM sqlite_master WHERE type='table' AND name='forums' ''')
if cur.fetchone()[0] == 0:
    set_up_merged_db(conn)

cur.execute(''' SELECT count(id) FROM forums ''')
if cur.fetchone()[0] == 0:
    url = "https://www.netmums.com/coffeehouse/"
    soup = get_soup(url)
    forums = soup.find_all("div", {"class": "left"})
    for forum in forums:
        if forum_link := forum.find('a', href=True):
            parsed = (forum_link['href'].strip(), forum_link.get_text().strip())
            sql = '''
                INSERT INTO forums(url,name)
                VALUES(?,?)
            '''
            cur.execute(sql, parsed)
conn.commit()

## Subforums

cur.execute(''' SELECT count(id) FROM subforums ''')
if cur.fetchone()[0] == 0:
    cur.execute(''' SELECT id, url FROM forums ''')
    rows = cur.fetchall()
    for row in rows:
        (forum_id, url) = row
        soup = get_soup(url)
        for subforum in soup.find_all("a", {"class": "cCatTopic"}):
            parsed = (subforum['href'].strip(), subforum.get_text().strip(),forum_id)
            sql = '''
                INSERT INTO subforums(url,name,forum_id)
                VALUES(?,?,?)
            '''
            cur.execute(sql, parsed)
conn.commit()

## Threads

cur.execute(''' SELECT count(id) FROM threads ''')
if cur.fetchone()[0] == 0:
    cur.execute(''' SELECT id, url FROM subforums ''')
    rows = cur.fetchall()
    for row in tqdm(rows):
        scrape_threads(cur, row)
        conn.commit()

## Posts
Create individual scraper files for each chunk of thread urls. Run with: ```nohup python3 netmums-group_xx.py > output_xx.txt &```

Posts can have different formatting in different situations, which leads to problems with parsing:
1. Must parse posts with citations to other posts differently than those without citations
2. Long urls are shortened when posted directly
3. WYSIWYG emojis must be translated from image to text
4. Skip the first post on each page after the first, it is duplicated from the last

This can be restart automatically. It deletes the last record collected in threads and starts rescraping that thread_id (if one exists).

cur.execute(''' SELECT max(id) FROM threads ''')
max_id = cur.fetchone()[0]

n_groups = 5
size = round(max_id / n_groups)

text = """#!/usr/bin/env python3
# coding: utf-8

## Imports

from netmums import *
from scraping import *
from pathlib import Path

## File Locations

p = Path.cwd()
path_parent = p.parents[0]
path_db_parent = str(path_parent / "database" / "netmums-merged.db")
path_db_child = str(path_parent / "database" / "netmums{0}.db")

## Connect to the database and create the tables.

### find max id in child db posts
conn = create_connection(path_db_child)
set_up_posts_db(conn)
cur = conn.cursor()
cur.execute(''' SELECT MAX(thread_id) FROM posts ''')
max_thread = cur.fetchone()[0]
if max_thread == None:
    first = {1}
else: # restart scraping from last complete thread id
    cur.execute(''' DELETE FROM posts WHERE thread_id=?''', (max_thread,))
    conn.commit()
    first = max_thread
conn.close()

### select rows from parent db threads
conn = create_connection(path_db_parent)
cur = conn.cursor()
cur.execute(''' SELECT id, url FROM threads WHERE id>=? AND id<={2} ''', (first,))
rows = cur.fetchall()
conn.close()

### connect back to child db
conn = create_connection(path_db_child)
set_up_posts_db(conn)

## Scrape threads

for row in rows:
    scrape_posts(conn, row)

conn.close()
"""


for g in range(n_groups):
    first = 1 + (size * g)
    if g == (n_groups - 1):
        last = max_id
    else:
        last = (size * (g + 1))
    if g + 1 < 10:
        num = "0{}".format(g + 1)
    else:
        num = "{}".format(g + 1)
    with open(path_groups.format(num), 'w') as w:
        w.write(text.format(num, first, last))

conn.close()

url = "https://www.netmums.com/coffeehouse/forum-faqs-1/forum-user-guide-534/1866241-becoming-ofsted-registered-nanny.html"

soup=get_soup(url)

canonical_link = soup.find("link", {"rel":"canonical"})
print(canonical_link["href"])

script = script.get_text()
script = json.loads(script)
posts = script["props"]["pageProps"]["initialReduxState"]["currentThread"]["currentThread"]["pagePosts"]

for post in posts:
    print(post)

cases = soup.find_all("div", {"class","md-topic_post"})

post = cases[0].find("div", {"class": "post_content"})
citations = post.find_all("span", {"itemprop": "citation"})
for i, cite in enumerate(citations):
    print(cite.find("span", {"itemprop": "url"}).get_text())
    print(cite.find("span", {"itemprop": "author"}).get_text())
    print(cite.find("span", {"itemprop": "text"}).get_text())
