==> scripts/netmums/2.2-Sentiment-Sentiment_Day.ipynb <==
# Generate sentiment and emotion by day

## Imports

import sqlite3
import pandas as pd
import numpy as np
# from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
from pathlib import Path
# from scraping import create_connection
# from tqdm.notebook import tqdm
# from math import floor
# import lemmatize as l
import dask.dataframe as dd
from dask.multiprocessing import get
import time
from torch import tensor
from scipy.special import softmax


## File Locations

p = Path.cwd()
path_parent = p.parents[1]
path_db = str(path_parent / "database" / "netmums-merged.db")

path_clean_data = path_parent / "clean_data" / "netmums"
path_text_pkl = str(path_clean_data / "daily_clean_text.pkl")
path_emote_text_pkl = str(path_clean_data / "daily_emote_clean_text_{}.pkl")
path_emote_processed_pkl = str(path_clean_data / "daily_emote_processed_{}.pkl")

## Get Daily Text
env=forum

sql = '''
    SELECT
        text.text_clean AS text_clean,
        s.name AS subforum_name,
        p.user_url AS user_url,
        p.date_created AS date_created
    FROM text
    LEFT JOIN posts AS p
    ON text.post_id = p.id
    LEFT JOIN threads AS t
    ON t.id=p.thread_id
    LEFT JOIN subforums AS s
    ON s.id=t.subforum_id
    LEFT JOIN forums AS f
    ON f.id=s.forum_id
    WHERE text.text_clean<>""
    AND p.user_url<>"Anonymous"
'''

conn = create_connection(path_db)
df = pd.read_sql_query(sql, conn)
conn.close()

ddf = dd.from_pandas(df, npartitions=200)
ddf['date_created'] = dd.to_datetime(ddf['date_created'])
ddf['day'] = ddf['date_created'].dt.date
ddf = ddf[['user_url', 'day', 'text_clean']].groupby(["user_url", "day"])["text_clean"].apply(lambda grp: ' '.join(grp), meta=('text', 'object'))
df = ddf.compute(scheduler='processes')

df = df.reset_index(drop=False)

df.to_pickle(path_text_pkl)

## Sentiment

df = pd.read_pickle(path_text_pkl)

analyzer = SentimentIntensityAnalyzer()

df['sentiment'] = df['text'].apply(lambda x: analyzer.polarity_scores(x))

df.to_pickle(path_text_pkl)

## Chunk Text
env=hf

from transformers import AutoModelForSequenceClassification, AutoTokenizer
path_model = '/home/mwh/cardiffnlp/twitter-roberta-base-emotion/'
model = AutoModelForSequenceClassification.from_pretrained(path_model)
tokenizer = AutoTokenizer.from_pretrained(path_model)

def split_text(text, max_size):
    """ 
    """
    # filter long words
    text = " ".join([word for word in text.split(" ") if len(word) <=25])
    # tokenize text
    encoded = tokenizer.encode(text)[1:-1]
    n_tokens = len(encoded)
    if n_tokens >= max_size:
        n_chunks = n_tokens // max_size + (n_tokens % max_size > 0) # round up
        chunk_size = n_tokens // n_chunks + (n_tokens % n_chunks > 0)
        return [[0] + encoded[i:i + chunk_size] + [2] for i in range(0, n_tokens, chunk_size)]
    else:
        return [[0] + encoded + [2]]

def make_tensor(encoded_list):
    output = {
        'input_ids': tensor([encoded_list]),
        'attention_mask': tensor([[1 for i in encoded_list]])
    }
    return output

def get_emotion_scores(encoded_tensor):
    output = model(**encoded_tensor)
    scores = output[0][0].detach().numpy()
    scores = softmax(scores)
    return list(scores)

text_df = pd.read_pickle(path_text_pkl)
more_than_one = text_df.groupby("user_url")['day'].count()
more_than_one = more_than_one[more_than_one > 1].reset_index(drop=False)
text_df = text_df.loc[text_df['user_url'].isin(more_than_one['user_url'])].reset_index(drop=True)

remaining = text_df.shape[0] - 1000000
half = remaining // 2
start = 1000000
ranges = [(1, start, start+half), (2, start+half, start+remaining)]

for i, start, end in ranges:
    start_time = time.time()
    df = text_df.iloc[start:end].copy()
    print("splitting text")
    ddf = dd.from_pandas(df, npartitions=200)
    ddf['encoded'] = ddf.apply(lambda x: split_text(x['text'], 500), axis=1, meta=list)
    df = ddf.compute(scheduler='processes')
    df.to_pickle(path_emote_text_pkl.format(i))
    split_time = time.time()
    print("time: {}".format((split_time - start_time) / 60))
    print("stacking encoded words")
    df = df.set_index(['user_url', 'day'])['encoded'].apply(pd.Series).stack().reset_index()
    df.columns = ['user_url', 'day', 'list_number', 'encoded']
    df = df[['user_url','day','encoded']]
    df.to_pickle(path_emote_text_pkl.format(i))
    stack_time = time.time()
    print("time: {}".format((stack_time - split_time) / 60))
    print("making tensors")
    ddf = dd.from_pandas(df, npartitions=200)
    ddf['encoded'] = ddf.apply(lambda x: make_tensor(x['encoded']), axis=1, meta=dict)
    df = ddf.compute(scheduler='processes')
    df.to_pickle(path_emote_text_pkl.format(i))
    tensor_time = time.time()
    print("time: {}".format((tensor_time - stack_time) / 60))
    print("get emotion scores")
    ddf = dd.from_pandas(df, npartitions=10)
    ddf['scores'] = ddf.apply(lambda x: get_emotion_scores(x['encoded']), axis=1, meta=list)
    df = ddf.compute(scheduler='processes')
    df = df[['user_url', 'day', 'scores']]
    df[['anger', 'joy', 'optimism', 'sadness']] = pd.DataFrame(df["scores"].to_list(), index=df.index)
    df.to_pickle(path_emote_processed_pkl.format(i))
    emotion_time = time.time()
    print("time: {}".format((emotion_time - tensor_time) / 60))
    print("total loop time: {}".format((emotion_time - start_time) / 60))

df.head()
