==> scripts/netmums/3.0-Topic_Models-Lemmatize_Text.ipynb <==
# Lemmatize Forum Text
Cleans the data for topic modelling


from lemmatize import *

group = "special-needs"
id_type = "thread_id"  # thread_id groups all messages in a thread
                       # together, message_id keeps them separated
per = 1
n_chunks = 4 # need at least 4 for all, all

process_data("special-needs", group, id_type, per, n_chunks)

process_data("being-a-mum", group, id_type, per, 4)

process_data("legal", group, id_type, per, 4)

process_data("education", group, id_type, per, 4)

process_data(forum="all", group="all", id_type="family_id", per=.25, n_chunks=4)
