==> scripts/netmums/5.2-Propensity-SN_Users.ipynb <==
# Generate Word Features


## Imports

# LDA
from gensim import corpora, models
from gensim.models import CoherenceModel, LdaModel, LdaMulticore
from gensim.models.callbacks import PerplexityMetric, ConvergenceMetric, CoherenceMetric
# Managing data
import pandas as pd
import numpy as np
import re
# DB connection
from scraping import create_connection
# Files & I/O
import pickle
import csv
import os
from pathlib import Path
from io import FileIO
# For logging
import logging
# Plotting
import matplotlib.pyplot as plt
# Random
import random
# Parallelizing
import dask.dataframe as dd
from dask.multiprocessing import get

## Functions

## File Locations

p = Path.cwd()
path_parent = p.parents[1]

# database
path_db = str(path_parent / "database" / "netmums-merged.db")
path_clean_data = path_parent / "clean_data" / "netmums"
# data to load
path_lemma_pkl = str(path_clean_data / "lemmatized_text_{0}_{1}_{2}.pkl")
path_corpus_pkl = str(path_clean_data / "corpus_{0}_{1}_{2}.pkl")
path_dictionary_gensim = str(path_clean_data / "dictionary_{0}_{1}_{2}.gensim")
# model saving
path_tune_models = str(path_clean_data / "lda_tune_{0}_{1}_{2}_{3}_{4}.gensim")
path_ntopic_models = str(path_clean_data / "lda_ntopics_{0}_{1}_{2}_{3}.gensim")
# path_coherence = str(path_parent / "clean_data" / "coherence_{}.csv")
path_log = str(path_clean_data / "logging_{0}_{1}_{2}_{3}.log")
path_log_iterations = str(path_clean_data / "logging_{0}_{1}_{2}_{3}.log")
# dominant topic
path_dom_topic = str(path_clean_data / "dominant_topic_{0}_{1}_{2}_{3}.csv")

# dataframes
path_topics_pkl = str(path_clean_data / "daily_topics_sn.pkl")
path_text_pkl = str(path_clean_data / "daily_clean_text.pkl")
path_topics = str(path_clean_data / "lda_topics_sn.csv")
# path_days_since_pkl = str(path_clean_data / "daily_days_since.pkl")
# path_subforums_pkl = str(path_clean_data / "daily_subforums.pkl")
# path_emote_pkl = str(path_clean_data / "daily_emote_processed_{}.pkl")
# path_joined_pkl = str(path_clean_data / "daily_joined_df.pkl")
# path_bigrams_pkl = str(path_corpus_pkl.format("all", "all", "daily_text_df"))

## Create Topic Model

forum = "special-needs"
group = "all"
id_type = "family_id"

dictionary = corpora.Dictionary.load(path_dictionary_gensim.format(forum, group, id_type))
corpus = pickle.load(open(path_corpus_pkl.format(forum, group, id_type), 'rb'))
lemmatized_text = pickle.load(open(path_lemma_pkl.format(forum, group, id_type), 'rb'))

import time

# 15 topics has the highest coherence
n_topics = 15
start = time.time()
lda = LdaModel(
    corpus=corpus,
    num_topics=n_topics,
    id2word=dictionary,
    random_state=238,
    alpha="auto",
    eta="auto"
)
print("model finished time: ", (time.time() - start) / 60)
coherence_model_lda = CoherenceModel(model=lda, texts=lemmatized_text, dictionary=dictionary, coherence='c_v')
print(coherence_model_lda.get_coherence())
print("working time: ", (time.time() - start) / 60)

lda.save(path_ntopic_models.format(forum, group, id_type, str(n_topics)))

## Daily text to corpus

import lemmatize_all as la

dictionary = corpora.Dictionary.load(path_dictionary_gensim.format(forum, group, id_type))
clean_text = pd.read_pickle(path_text_pkl)
clean_text.columns = ['user_url', 'day', 'text_clean', 'sentiment']

sn_users_sql = """
    SELECT DISTINCT p.user_url AS user_url
    FROM posts AS p
    LEFT JOIN threads AS t
    ON t.id=p.thread_id
    LEFT JOIN subforums AS s
    ON s.id=t.subforum_id
    WHERE s.forum_id=24
"""
conn = create_connection(path_db)
sn_user_df = pd.read_sql_query(sn_users_sql, conn)
conn.close()

clean_text = pd.merge(clean_text, sn_user_df, on="user_url", how="inner")

%%time
df_bigrams, corpus = la.make_corpus(clean_text, dictionary, 4)

df_bigrams.to_pickle(path_corpus_pkl.format("special-needs", "all", "daily_text_df"))

pickle.dump(corpus, open(path_corpus_pkl.format("special-needs", "all", "daily_text"), 'wb'))

# corpus = pickle.load(open(path_corpus_pkl.format("special-needs", "all", "daily_text"), 'rb'))
# len(corpus)

## Daily text to topics

n_topics = 15
lda = LdaModel.load(path_ntopic_models.format(forum, group, id_type, str(n_topics)))

from gensim.matutils import corpus2csc
import numpy as np

all_topics = lda.get_document_topics(corpus, minimum_probability=0.0)
all_topics_csr = corpus2csc(all_topics)
all_topics_numpy = all_topics_csr.T.toarray()
all_topics_df = pd.DataFrame(all_topics_numpy)
all_topics_df.head()

column_names = ["topic_{}".format(i) for i in range(n_topics)]
all_topics_df.columns = column_names
all_topics_df.to_pickle(path_topics_pkl)

df_bigrams.info()

all_topics_df.info()

all_topics_df = pd.concat([df_bigrams.reset_index(drop=True), all_topics_df.reset_index(drop=True)], axis=1)

all_topics_df = all_topics_df.drop('bigrams', axis=1)

all_topics_df.to_pickle(path_topics_pkl)
all_topics_df.to_csv(str(path_clean_data / "daily_topics_sn.csv"), index=False)

all_topics_df = pd.read_pickle(path_topics_pkl)

## Get Topic Keywords

def write_list(fn, results):
    with open(fn, 'a') as f:
        writer = csv.writer(f) 
        writer.writerow(results)

n_topics = 15
lda = LdaModel.load(path_ntopic_models.format(forum, group, id_type, str(n_topics)))

n_words = 20
topics = lda.print_topics(num_topics=n_topics, num_words=n_words)
for topic in topics:
    write_list(path_topics, [topic[0], topic[1]])

## Save Dominant Topic and Text

clean_text.head()

all_topics_df.head()

all_topics_df.columns

all_topics_df['dominant_topic'] = all_topics_df[['topic_0', 'topic_1', 'topic_2', 'topic_3',
       'topic_4', 'topic_5', 'topic_6', 'topic_7', 'topic_8', 'topic_9',
       'topic_10', 'topic_11', 'topic_12', 'topic_13', 'topic_14']].idxmax(axis=1)

all_topics_df['dominant_topic'].value_counts()

all_topics_text = pd.merge(all_topics_df, clean_text[['user_url','day','text_clean']], on=['user_url','day'], how="inner")

all_topics_text.to_csv(str(path_clean_data / "sn_topics_and_text.csv"), index=False)
